+++ Aurora/include/utils/circ_buf.h
@@ -29,7 +29,7 @@ extern void circ_buf_free (circ_buf_t *cbuf);
 extern size_t circular_buf_size(circ_buf_t *cbuf);
 extern size_t circular_buf_capacity(circ_buf_t *cbuf);
 extern void circular_buf_put(circ_buf_t *cbuf,uint8_t data);
-extern int circular_buf_put2(circ_buf_t *cbuf, int data);
+extern int circular_buf_put2(circ_buf_t *cbuf, uint8_t data);
 extern int circular_buf_get(circ_buf_t *cbuf, uint8_t *data);
 extern int circular_buf_get2(circ_buf_t *cbuf, int *data);
 extern bool circular_buf_empty(circ_buf_t *cbuf);
